Add metadata for elcipse compiler in Windows
Add the subfolder .metadata that eclipse creates for the workspace. This is for Windows version of eclipse Kepler.

If you happen to have a WIndows machine with eclipse Kepler + CDT + MinGW, you can unzip this file into the root directory and open this workspacein eclipse. Ya can then compile your project in or x86 architecture.